feat(unsetcompact): adds special [] syntax to unsetCompact
Using an empty array syntax in a dot notation string allows recursively unsetting values by key in
all elements of an array
