+++ src/Inputs/index.ts
@@ -4,6 +4,9 @@ import { UseInertiaInputProps } from '../useInertiaInput'
 
 export { default as Input } from './Input'
 export { default as Submit } from './Submit'
+export { default as NestedFields, type NestedFieldsProps } from './NestedFields'
+export { default as DynamicInputs }  from './DynamicInputs'
+export { default as useDynamicInputs, type DynamicInputsProps } from './useDynamicInputs'
 
 export type InputConflicts = 'name'|'onChange'|'onBlur'|'onFocus'|'value'|'defaultValue'
 export interface BaseFormInputProps<T, TForm extends NestedObject = NestedObject>
