+++ src/Inputs/useDynamicInputs.ts
@@ -1,5 +1,5 @@
 import { useCallback } from 'react'
-import { useForm, useFormMeta } from './Form'
+import { useForm, useFormMeta } from '../Form'
 import { get, set } from 'lodash'
 import { useNestedAttribute } from './NestedFields'
