+++ src/index.ts
@@ -1,6 +1,5 @@
 export { default as useInertiaForm, type UseInertiaFormProps, type NestedObject } from './useInertiaForm'
 export { default as useInertiaInput, type UseInertiaInputProps } from './useInertiaInput'
-export { default as useDynamicInputs } from './useDynamicInputs'
 export {
 	Form,
 	useForm,
@@ -8,6 +7,12 @@ export {
 	type UseFormProps,
 	type FormProps,
 } from './Form'
-export { default as NestedFields, type NestedFieldsProps } from './NestedFields'
-export { default as DynamicInputs, type DynamicInputsProps } from './Inputs/DynamicInputs'
-export { Input, Submit } from './Inputs'
+export {
+	Input,
+	Submit,
+	NestedFields,
+	useDynamicInputs,
+	type NestedFieldsProps,
+	DynamicInputs,
+	type DynamicInputsProps,
+} from './Inputs'
