+++ tests/utils/unsetCompact.test.ts
@@ -1,5 +1,5 @@
-import { NestedObject } from '../src/useInertiaForm'
-import { coerceArray, fillEmptyValues, renameObjectWithAttributes, stripAttributes, unsetCompact } from '../src/utils'
+import { NestedObject } from '../../src/useInertiaForm'
+import { unsetCompact } from '../../src/utils'
 
 const nestedData: NestedObject = {
 	one: 'one',
@@ -43,7 +43,7 @@ describe('unsetCompact', () => {
 		})
 	})
 
-	it('should reorder arrays making all elements sequential and removing empty array elements', () => {
+	it('should delete by array index, filtering out empty array elements', () => {
 		const data = structuredClone(nestedData)
 
 		unsetCompact(data, 'two.four[0]')
@@ -67,9 +67,9 @@ describe('unsetCompact', () => {
 	})
 
 	describe('recursively unsets array elements by key with empty array brackets', () => {
-		const data = structuredClone(nestedData)
-
 		it('unsets all instances of a key', () => {
+			const data = structuredClone(nestedData)
+
 			unsetCompact(data, 'two.four[].five')
 			expect(data).toEqual({
 				one: 'one',
@@ -91,15 +91,17 @@ describe('unsetCompact', () => {
 		})
 
 		it('works with nested array objects', () => {
+			const data = structuredClone(nestedData)
+
 			unsetCompact(data, 'two.four[].ten[].twelve')
 			expect(data).toEqual({
 				one: 'one',
 				two: {
 					three: 'three',
 					four: [
-						{ six: 'six' },
+						{ five: 'five', six: 'six' },
 						{ seven: 'seven' },
-						{ six: 'nine', ten: [
+						{ five: 'eight', six: 'nine', ten: [
 							{ eleven: 'eleven' },
 							{ eleven: 'eleven', thirteen: 'thirteen' },
 						] },
@@ -112,17 +114,19 @@ describe('unsetCompact', () => {
 		})
 
 		it('works when an element is specified after an empty bracket', () => {
-			unsetCompact(data, 'two.four[].ten[1].thirteen')
+			const data = structuredClone(nestedData)
+
+			unsetCompact(data, 'two.four[].ten[1].eleven')
 			expect(data).toEqual({
 				one: 'one',
 				two: {
 					three: 'three',
 					four: [
-						{ six: 'six' },
+						{ five: 'five', six: 'six' },
 						{ seven: 'seven' },
-						{ six: 'nine', ten: [
-							{ eleven: 'eleven' },
-							{ eleven: 'eleven' },
+						{ five: 'eight', six: 'nine', ten: [
+							{ eleven: 'eleven', twelve: 'twelve' },
+							{ thirteen: 'thirteen' },
 						] },
 					],
 					last: {
@@ -133,15 +137,20 @@ describe('unsetCompact', () => {
 		})
 
 		it('works when an empty bracket is specified after an element ', () => {
+			const data = structuredClone(nestedData)
+
 			unsetCompact(data, 'two.four[2].ten[].eleven')
 			expect(data).toEqual({
 				one: 'one',
 				two: {
 					three: 'three',
 					four: [
-						{ six: 'six' },
+						{ five: 'five', six: 'six' },
 						{ seven: 'seven' },
-						{ six: 'nine', ten: [] },
+						{ five: 'eight', six: 'nine', ten: [
+							{ twelve: 'twelve' },
+							{ thirteen: 'thirteen' },
+						] },
 					],
 					last: {
 						just: 'testing',
@@ -150,89 +159,5 @@ describe('unsetCompact', () => {
 			})
 
 		})
-	})
-})
-
-describe('fillEmptyValues', () => {
-	it('should replace undefined or null values with empty strings', () => {
-		const sanitized = fillEmptyValues({
-			one: 'one',
-			two: undefined,
-			three: null,
-			nested: {
-				four: undefined,
-				five: null,
-			},
-		})
-
-		expect(sanitized).toMatchObject({
-			one: 'one',
-			two: '',
-			three: '',
-			nested: {
-				four: '',
-				five: '',
-			},
-		})
-	})
-})
-
-describe('stripAttributes', () => {
-	it('should remove _attributes from dot notation strings', () => {
-		const key = 'user.person_attributes.name'
-		expect(stripAttributes(key)).toEqual('user.person.name')
-	})
-
-	it('should not remove the word attributes without the _ or not at the end of a segment', () => {
-		const key1 = 'user.attributes'
-		const key2 = 'user.manager_attributes_sorted'
-		expect(stripAttributes(key1)).toEqual(key1)
-		expect(stripAttributes(key2)).toEqual(key2)
-	})
-})
-
-const mockFormData = {
-	person: {
-		first_name: 'Something',
-		user: {
-			username: 'something',
-		},
-		contact: {
-			emails: [
-				{ email: 'something@email.com' },
-			],
-		},
-	},
-}
-
-describe('renameObjectWithAttributes', () => {
-	it('should append attributes to keys deeper than the first level which contain another object', () => {
-		const data = renameObjectWithAttributes(mockFormData)
-
-		expect(data).toMatchObject({
-			person: {
-				first_name: 'Something',
-				user_attributes: {
-					username: 'something',
-				},
-				contact_attributes: {
-					emails_attributes: [
-						{ email: 'something@email.com' },
-					],
-				},
-			},
-		})
-	})
-})
-
-describe('coerceArray', () => {
-	it('should return an array regardless of input', () => {
-		const nonArray = 'hello'
-		const array = ['one', 'two', 'three']
-		const coerced1 = coerceArray(nonArray)
-		const coerced2 = coerceArray(array)
-
-		expect(coerced1).toEqual([nonArray])
-		expect(coerced2).toEqual(array)
 	})
 })
